♻ refactor: rename and move dbConnection
